fix errors for bank page and add merchant code
