+++ Shichi/CustomUser/urls.py
@@ -2,7 +2,7 @@ from django.urls import path, include
 from .views import CustomUserViewSet, CreateCustomUser, MyCustomUser, ShowCustomUser, UpdateCustomUser, DeleteCustomUser
 from .views import PasswordResetView, PasswordResetConfirmView
 from .views import WalletListView, WalletCreateView, WalletUpdateView, WalletDeleteView
-from .views import go_to_gateway_view, callback_gateway_view
+
 from rest_framework_simplejwt.views import (
     TokenObtainPairView,
     TokenRefreshView,
@@ -15,7 +15,6 @@ from django.contrib.auth.views import (
 )
 
 app_name = "user"
-from azbankgateways.urls import az_bank_gateways_urls
 
 urlpatterns = [
     path('myshow/', MyCustomUser.as_view(), name='myshow'),
@@ -35,9 +34,5 @@ urlpatterns = [
     path('WalletUpdate/<int:pk>/', WalletUpdateView.as_view(), name='WalletUpdate'),
     # path('WalletDelete/<int:pk>/', DeleteCustomUser.as_view(), name='WalletDelete'),
     # path('password-reset-complete/',PasswordResetCompleteView.as_view(),name='password_reset_complete'),
-    
-    path('bankgateways/', az_bank_gateways_urls()),
-    path('go-to-gateway/', go_to_gateway_view),
-    path('callback-gateway/', callback_gateway_view),
 ]
