+++ Shichi/Payments/views.py
@@ -1,0 +1,48 @@
+from django.shortcuts import render
+import logging
+from django.urls import reverse
+from azbankgateways import bankfactories, models as bank_models, default_settings as settings
+from azbankgateways.exceptions import AZBankGatewaysException
+from django.http import HttpResponse, Http404
+
+# Create your views here.
+def go_to_gateway_view(request):
+    amount = 10000
+    # تنظیم شماره موبایل کاربر از هر جایی که مد نظر است
+    user_mobile_number = '+989112521234'  # اختیاری
+    factory = bankfactories.BankFactory()
+    try:
+        # bank = factory.auto_create() # or 
+        bank = factory.create(bank_models.BankType.ZARINPAL) 
+        bank.set_request(request)
+        bank.set_amount(amount)
+        bank.set_client_callback_url(reverse('callback-gateway'))  # reverse...
+        bank.set_mobile_number(user_mobile_number)  # اختیاری
+        bank_record = bank.ready()
+        return bank.redirect_gateway()
+    except AZBankGatewaysException as e:
+        logging.critical(e)
+        # TODO: redirect to failed page.
+        raise e
+
+
+def callback_gateway_view(request):
+    tracking_code = request.GET.get(settings.TRACKING_CODE_QUERY_PARAM, None)
+    if not tracking_code:
+        logging.debug("این لینک معتبر نیست.")
+        raise Http404
+
+    try:
+        bank_record = bank_models.Bank.objects.get(tracking_code=tracking_code)
+    except bank_models.Bank.DoesNotExist:
+        logging.debug("این لینک معتبر نیست.")
+        raise Http404
+
+    # در این قسمت باید از طریق داده هایی که در بانک رکورد وجود دارد، رکورد متناظر یا هر اقدام مقتضی دیگر را انجام دهیم
+    if bank_record.is_success:
+        # پرداخت با موفقیت انجام پذیرفته است و بانک تایید کرده است.
+        # می توانید کاربر را به صفحه نتیجه هدایت کنید یا نتیجه را نمایش دهید.
+        return HttpResponse("پرداخت با موفقیت انجام شد.")
+
+    # پرداخت موفق نبوده است. اگر پول کم شده است ظرف مدت ۴۸ ساعت پول به حساب شما بازخواهد گشت.
+    return HttpResponse("پرداخت با شکست مواجه شده است. اگر پول کم شده است ظرف مدت ۴۸ ساعت پول به حساب شما بازخواهد گشت.")
